% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concat_avgSP_timeseries}
\alias{concat_avgSP_timeseries}
\title{Concatenate time series of average profiles}
\usage{
concat_avgSP_timeseries(avgSP1, avgSP2)
}
\arguments{
\item{avgSP1}{old time series of average profiles as returned by \link{averageSPalongSeason}}

\item{avgSP2}{new time series of average profiles as returned by \link{averageSPalongSeason}}
}
\description{
This is useful in operations to update a time series that was computed in the past
with a newly computed average time series. The routine merges all entries with duplicated
entries (read dates) being taken from \code{avgSP2}.
}
\seealso{
\link{averageSPalongSeason}
}
\author{
fherla
}
