% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{densityDistance}
\alias{densityDistance}
\title{Difference in layer density}
\usage{
densityDistance(density1, density2, normalize = FALSE, absDist = TRUE)
}
\arguments{
\item{density1}{numeric density values (1D array)}

\item{density2}{numeric density values (1D array)}

\item{normalize}{Should result be normalized? boolean, default False.}

\item{absDist}{Interested in absolute distance? default True.}
}
\value{
numeric density distance
}
\description{
Calculate the difference (i.e. distance) in layer density
}
\author{
pbillecocq
}
