% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceSP.R
\name{distanceSP}
\alias{distanceSP}
\title{Wrapper for dtwSP and simSP}
\usage{
distanceSP(query, ref, ...)
}
\arguments{
\item{query}{The query snowprofile object (will be warped onto \code{ref})}

\item{ref}{The reference snowprofile object (will \emph{not} be warped)}

\item{...}{passed on to \link{dtwSP}}
}
\description{
Calculate the distance between two snowprofile objects by
}
\details{
\enumerate{
\item Matching their layers and aligning them (i.e., warp one profile onto the other one)
\item Assessing the similarity of the aligned profiles based on avalanche hazard relevant characteristics
\item Convert the similarity score into a distance value between \verb{[0, 1]}
}

This procedure is useful for clustering and aggregating tasks, given a set of multiple profiles.
}
\seealso{
\link{dtwSP}, \link{simSP}, \link{medoidSP}
}
\author{
fherla
}
