% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flipLayers}
\alias{flipLayers}
\title{Flip snow profile layers top down}
\usage{
flipLayers(x)
}
\arguments{
\item{x}{snowprofile or snowprofileLayers object with layers to be flipped}
}
\value{
same object with layers dataframe flipped upside down
}
\description{
Flip snow profile layers top down
}
\note{
only do that with a specific reason (better, don"t do it!), as all functions with snowprofile objects are
designed to have the layers increase in height.
}
