% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainSimilarityMatrices.R
\name{grainSimilarity_evaluate}
\alias{grainSimilarity_evaluate}
\title{Grain type similarity matrix for evaluation purposes}
\usage{
grainSimilarity_evaluate(triag = TRUE)
}
\arguments{
\item{triag}{Return a triangular matrix (TRUE, default) or a symmetric matrix (FALSE)}
}
\value{
data.frame, either triangular or symmetric
}
\description{
Similar to \link{grainSimilarity_align}, but designed for assessing the similarity between snow profiles based
on avalanche hazard relevant characteristics. To be used in combination with \link{simSP}.
}
\examples{

simMat <- grainSimilarity_evaluate()
print(simMat)

}
\author{
fherla
}
