% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hardnessDistance}
\alias{hardnessDistance}
\title{Difference in Hand Hardness}
\usage{
hardnessDistance(hardness1, hardness2, normalize = FALSE, absDist = TRUE)
}
\arguments{
\item{hardness1}{character or numeric hand hardness value (1D array)}

\item{hardness2}{character or numeric hand hardness value (1D array)}

\item{normalize}{Should result be normalized? boolean, default False.}

\item{absDist}{Interested in absolute distance? default True.}
}
\value{
numeric Hand Hardness Distance
}
\description{
Calculate the difference (i.e. distance) in hand hardness
}
\author{
fherla
}
