% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveAlignment.R
\name{interactiveAlignment}
\alias{interactiveAlignment}
\title{Run interactive alignment app}
\usage{
interactiveAlignment(query = NaN, ref = NaN)
}
\arguments{
\item{query}{an optional query snowprofile}

\item{ref}{an optional reference snowprofile}
}
\value{
An interactive session will be started
}
\description{
This app allows to interactively explore the alignment of two snowprofiles, which are either
given as input to this function, or are uploaded to the app interactively as caaml files.
Example profiles are also provided in the app.
}
\examples{
if (FALSE){  # this example won't be started in package tests.

## start app and choose profiles from within the app:
interactiveAlignment()

## start app with package internal profile data (from `sarp.snowprofile`):
interactiveAlignment(query = SPpairs$A_modeled, ref = SPpairs$A_manual)

}

}
\author{
fherla
}
