% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainSimilarityMatrices.R
\name{layerWeightingMat}
\alias{layerWeightingMat}
\title{Weighting scheme for preferential layer matching}
\usage{
layerWeightingMat(triag = TRUE)
}
\arguments{
\item{triag}{Return a triangular matrix (TRUE, default) or a symmetric matrix (FALSE)}
}
\value{
data.frame, either triangular or symmetric
}
\description{
A matrix, of the same form as \link{grainSimilarity_align}, but containing weighting coefficients for preferential layer
matching based on the grain types of the layers.
}
\examples{

weightsMat <- layerWeightingMat()
print(weightsMat)

}
\author{
fherla
}
