% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_with_tolerance}
\alias{match_with_tolerance}
\title{Match with numeric tolerance}
\usage{
match_with_tolerance(x, y, d = 2)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{d}{numeric tolerance in form of digits}
}
\value{
boolean vector equivalently to \link{match}
}
\description{
Match with numeric tolerance
}
