% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ogsDistance}
\alias{ogsDistance}
\title{Difference in layer ogs}
\usage{
ogsDistance(ogs1, ogs2, normalize = FALSE, absDist = TRUE)
}
\arguments{
\item{ogs1}{numeric ogs values (1D array)}

\item{ogs2}{numeric ogs values (1D array)}

\item{normalize}{Should result be normalized? boolean, default False.}

\item{absDist}{Interested in absolute distance? default True.}
}
\value{
numeric ogs distance
}
\description{
Calculate the difference (i.e. distance) in layer ogs
}
\author{
pbillecocq
}
