% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{puDistance}
\alias{puDistance}
\title{Difference in layer stability p_unstable}
\usage{
puDistance(pu1, pu2)
}
\arguments{
\item{pu1}{numeric p_unstable values (1D array)  (within \verb{[0, 1]})}

\item{pu2}{numeric p_unstable values (1D array)  (within \verb{[0, 1]})}
}
\value{
numeric p_unstable distance
}
\description{
Calculate the difference (i.e. distance) in p_unstable
}
\author{
fherla
}
