% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_conceptually_similar_gtypes}
\alias{return_conceptually_similar_gtypes}
\title{Return conceptually similar grain types}
\usage{
return_conceptually_similar_gtypes(gt)
}
\arguments{
\item{gt}{a single gtype}
}
\value{
a character vector of similar gtypes
}
\description{
Note, use this function with care. It's a brief helper function for specific usage, not generally applicable!
It is, however, sometimes useful for backtracking layers, see \link{backtrackLayers}.
}
\examples{
return_conceptually_similar_gtypes("SH")
return_conceptually_similar_gtypes("MFcr")
return_conceptually_similar_gtypes("RG")
}
