% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaleSnowHeight}
\alias{scaleSnowHeight}
\title{Scale total height of a snow profile}
\usage{
scaleSnowHeight(query, ref = NA, height = NA)
}
\arguments{
\item{query}{the query snow profile (whose height vector will be scaled)}

\item{ref}{the reference snow profile (whose total snow height will be used as the reference height for the scaling)}

\item{height}{an optional reference height that can be given instead of the query profile}
}
\value{
query profile with scaled height vector
}
\description{
Scale the snow height of a snow profile either (1) based on another profile, or (2) based on a provided (predetermined) snow height.
This function can therefore be used to scale two snow profiles to an identical snow height by scaling the height vector of the (query) profile
against the height vector of the (reference) profile.
}
\author{
fherla
}
