% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SPgroup2.R
\docType{data}
\name{SPgroup2}
\alias{SPgroup2}
\title{Additional example set of snow profiles}
\format{
A \link[sarp.snowprofile:snowprofileSet]{sarp.snowprofile::snowprofileSet}
}
\usage{
SPgroup2
}
\description{
Additional example set of snow profiles. The main difference to the example data set \link[sarp.snowprofile:SPgroup]{sarp.snowprofile::SPgroup} is that \code{SPgroup2} contains various different stability indices.
}
\examples{

plot(SPgroup2, SortMethod = "unsorted")

}
\seealso{
\link[sarp.snowprofile:SPgroup]{sarp.snowprofile::SPgroup}
}
\keyword{datasets}
