% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SPspacetime.R
\docType{data}
\name{SPspacetime}
\alias{SPspacetime}
\title{Additional example set of snow profiles}
\format{
A \link[sarp.snowprofile]{snowprofileSet}
}
\usage{
SPspacetime
}
\description{
Additional example set of 4 spatially distributed snow profiles for 5 consecutive days, also containing different stability indices.
}
\examples{

plot(SPspacetime, SortMethod = "elev")

}
\seealso{
\link{SPgroup2}
}
\keyword{datasets}
