% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-swissSimilarityMatrix.R
\docType{data}
\name{swissSimilarityMatrix}
\alias{swissSimilarityMatrix}
\title{Similarity Matrix of Snow Grain Types}
\format{
A data.frame
}
\usage{
swissSimilarityMatrix
}
\description{
as defined by Lehning et al (2001). A similarity of 1 represents identity, 0 represents total dissimilarity.
}
\examples{

print(swissSimilarityMatrix)

}
\keyword{Grain}
\keyword{Similarity}
\keyword{Type}
